chore: add dummy provider and move provider to own folder
