+++ packages/cron/src/lib/email/providers/dummy.js
@@ -1,0 +1,8 @@
+/**
+ * An email provider that does nothing.
+ */
+export default class DummyEmailProvider {
+  async sendEmail () {
+    return Promise.resolve(`${Math.random()}`)
+  }
+}
